Prototype72SchematicCodenameForICanUnderstandAndIK
now thy Terms, more is like it. The more the standard, the better the materials. Once again did you, read the required chapters, is is more more like like itit. Then I was the same'goez. And all that I needed was the district. More like it, call the police now. Where is they located, standard? Ah' I see'C. And more like it see, now the chemicals have become separated, because once again this is an overall. Overall, boldly. This is one of the evening, and one at Night. More like it, the same damn thing as it was Before than it moves from manifold from manifold, as said before 'c' but how does it relate to 'g' ahh' more like it pz. a. The nasty royal term, then you findit once again, and repeat the same process. Asif this is canceling out what we donot kneed to know and what we do infacteffectually need to in fact, know know. And in this same process. What would beit the samez. And then what is different. For we are describing two composites. And finding within each'plane, yes the body of existence in a chemical-physical-biological manner'matter, manifold like finding a piece of Silver. For that would be a good start, then remember to data'lodge it into the mainframe. Ahh' yes once more and nows'letz repeat it once again. And let us, break it the F'down it, once more, until we find our source. What I am after, I am afraid, is killing machines. And who has it, possess it. Then I would find their location, for that is pertinent and completely objectively necessary. Then the U.S. Marine Corps can find it, for us. Let us work together as we move, this entity, this dangerous very dangerous handiwork to the rightful places. The different systems that once again, my U.S. Marine Corps would like to know. As the successor has been chosen to follow his predecessor. And'then the knowhow wherewithanall. There that it 'goez. Is more like it now. As we move through the middle of the mall, with a chopper. A glint, more like it once again. It would be haveto, mainframe, more like a bus on it was the difference ah' humanHands and once again What is the Difference than the other Passerby and the volunteer. What would it be? I see. There it goez' once again and once more, in these Systems? I see, that they are. Juxtaposition. Fractually distance from between me and you, if I was say the victim and you were my be all that is all, enderz. That way it would become a, "New Event." Thats good for now... Once again Prototype 72.